fastapi_template/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic[email]==1.8.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pytest==6.2.5
httpx==0.19.0
pillow==9.5.0
python-dotenv==1.0.0
aiofiles==0.8.0
numpy==1.24.3
torch==1.13.1
torchvision==0.14.1